Mediator artifactId change
