--- components/extensions/org.wso2.carbon.mediator.predict/src/main/resources/META-INF/services/org.apache.synapse.config.xml.MediatorSerializer
@@ -1,1 +1,0 @@
-org.wso2.carbon.mediator.predict.xml.PredictMediatorSerializer
