+++ components/extensions/org.wso2.carbon.ml.mediator.predict.ui/src/main/resources/org/wso2/carbon/ml/mediator/predict/ui/i18n/JSResources.properties
@@ -1,0 +1,1 @@
+mediator.predict.validInputmsg=Required fields can not be left blank
