+++ components/extensions/org.wso2.carbon.ml.mediator.predict.ui/src/main/resources/org/wso2/carbon/ml/mediator/predict/ui/i18n/Resources.properties
@@ -1,0 +1,16 @@
+predict.mediator=Predict Mediator
+mediator.predict.header=Predict Mediator
+mediator.predict.model.name=Model Storage Location
+mediator.predict.load.model=Load Model
+mediator.predict.error.msg=An error occurred while processing the request
+mediator.predict.model.not.found.msg=Model not found
+mediator.predict.invalid.model.msg=Please enter a valid model storage location
+mediator.predict.model.loaded.success=Model loaded successfully
+mediator.predict.features=Features
+mediator.predict.feature.name=Feature Name
+mediator.predict.expression=Expression
+mediator.predict.namespaces=Namespaces
+mediator.predict.prediction.output=Prediction Output
+mediator.predict.prediction.property=Property
+
+
