+++ components/extensions/org.wso2.carbon.ml.mediator.predict/src/main/resources/META-INF/services/org.apache.synapse.config.xml.MediatorSerializer
@@ -1,0 +1,1 @@
+org.wso2.carbon.ml.mediator.predict.xml.PredictMediatorSerializer
