+++ components/extensions/org.wso2.carbon.ml.mediator.predict/src/test/resources/test-model
