+++ components/extensions/pom.xml
@@ -29,7 +29,7 @@
     <packaging>pom</packaging>
     <name>WSO2 Machine Learner Extensions</name>
     <modules>
-        <module>org.wso2.carbon.mediator.predict</module>
-        <module>org.wso2.carbon.mediator.predict.ui</module>
+        <module>org.wso2.carbon.ml.mediator.predict</module>
+        <module>org.wso2.carbon.ml.mediator.predict.ui</module>
     </modules>
 </project>
