--- features/extensions/org.wso2.carbon.mediator.predict.feature/pom.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
- ~ Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- ~
- ~ Licensed under the Apache License, Version 2.0 (the "License");
- ~ you may not use this file except in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~      http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing, software
- ~ distributed under the License is distributed on an "AS IS" BASIS,
- ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- ~ See the License for the specific language governing permissions and
- ~ limitations under the License.
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <parent>
-        <groupId>org.wso2.carbon.ml</groupId>
-        <artifactId>ml-extension-features</artifactId>
-        <version>1.0.1-SNAPSHOT</version>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.mediator.predict.feature</artifactId>
-    <packaging>pom</packaging>
-    <name>WSO2 Carbon - Predict Mediator Aggregate Feature</name>
-    <url>http://wso2.org</url>
-    <description>This feature contains the bundles required for Predict Mediator functionality</description>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.wso2.carbon.ml</groupId>
-            <artifactId>org.wso2.carbon.mediator.predict.server.feature</artifactId>
-            <type>zip</type>
-	    </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.ml</groupId>
-            <artifactId>org.wso2.carbon.mediator.predict.ui.feature</artifactId>
-            <type>zip</type>
-	    </dependency>
-    </dependencies>
-    
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.wso2.maven</groupId>
-                <artifactId>carbon-p2-plugin</artifactId>
-                <version>${carbon.p2.plugin.version}</version>
-                <executions>
-                    <execution>
-                        <id>4-p2-feature-generation</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>p2-feature-gen</goal>
-                        </goals>
-                        <configuration>
-                            <id>org.wso2.carbon.mediator.predict</id>
-                            <propertiesFile>../../ml/feature.properties</propertiesFile>
-                            <importFeatures>
-                                <importFeatureDef>org.wso2.carbon.mediation.initializer</importFeatureDef>
-                            </importFeatures>
-                            <includedFeatures>
-                                <includedFeatureDef>org.wso2.carbon.ml:org.wso2.carbon.mediator.predict.server.feature</includedFeatureDef>
-                                <includedFeatureDef>org.wso2.carbon.ml:org.wso2.carbon.mediator.predict.ui.feature</includedFeatureDef>
-                            </includedFeatures>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
