+++ features/extensions/pom.xml
@@ -31,9 +31,9 @@
     <url>http://wso2.org</url>
 
     <modules>
-        <module>org.wso2.carbon.mediator.predict.server.feature</module>
-	    <module>org.wso2.carbon.mediator.predict.ui.feature</module>
-        <module>org.wso2.carbon.mediator.predict.feature</module>
+        <module>org.wso2.carbon.ml.mediator.predict.server.feature</module>
+	    <module>org.wso2.carbon.ml.mediator.predict.ui.feature</module>
+        <module>org.wso2.carbon.ml.mediator.predict.feature</module>
     </modules>
 </project>
